requirements.txt
groq>=0.11.0
feedparser>=6.0.11
python-dotenv>=1.0.1
